% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{cb.contains}
\alias{cb.contains}
\title{Does vector x violate cb with criterion L}
\usage{
cb.contains(x, cb.low, cb.high, L)
}
\arguments{
\item{x}{[M,1] numeric, A data vector}

\item{cb.low}{(1,m) numeric vector, lower CI limit for each of the dimensions}

\item{cb.high}{(1,m) numeric vector, upper CI limit for each of the dimensions}

\item{L}{[1,1] integer, Max number of outlier dimensions for a data row that is counted as within MVCI}
}
\value{
logical, TODO
}
\description{
TODO: a lot like ci_contains()
}
