% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{cb.envelope}
\alias{cb.envelope}
\title{Data envelope and related statistics}
\usage{
cb.envelope(data)
}
\arguments{
\item{data}{[N,M] numeric matrix, N samples, M variables/dimensions.}
}
\value{
A list with elements:
\item{envelope}{[2,M] matrix, lower (row 1) and upper (row 2) envelopes}
\item{sum.width}{[1,1] numeric, Envelope width (sum of widths over the M dimensions)}
\item{mean.width}{[1,1] numeric, Mean envelope width (mean of widths over the M dimensions)}
\item{min/max}{you get it...}
}
\description{
Data envelope and related statistics
}
