% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{cb.violations}
\alias{cb.violations}
\title{Count of cb violations in vector x}
\usage{
cb.violations(x, cb.low, cb.high)
}
\arguments{
\item{x}{[M,1] numeric, A data vector}

\item{cb.low}{(1,m) numeric vector, lower CI limit for each of the dimensions}

\item{cb.high}{(1,m) numeric vector, upper CI limit for each of the dimensions}
}
\value{
integer, Number of values in x that are outside cb. TODO: correct?
}
\description{
Returns the number of points in which a vector x lies outside cb
}
