% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{cb_area}
\alias{cb_area}
\title{Width of the mvci}
\usage{
cb_area(cb)
}
\arguments{
\item{cb}{list, mvci specifications produced by one of the mvci functions}
}
\value{
[1,1] numeric, Width of the mvci (the MWE target cost criterion)
}
\description{

}
