% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{cb_quantile}
\alias{cb_quantile}
\title{Quantile based naive confidence areas}
\usage{
cb_quantile(dmat, k, quantile.fun = min_width_quantile)
}
\arguments{
\item{dmat}{[N,M] numeric matrix, N samples, M variables.}

\item{k}{[1,1] integer, Number of observations to exclude}

\item{quantile.fun}{function, A function to compute the 1D quantiles with,
defaults to min_width_quantile()
Should accept two arguments: a data vector x and number of observations to
exclude k. See min_width_quantile() and central_quantile() for examples.}
}
\value{
A list with elements:
\item{$cb:}{[2,M] numeric array containing the lower and upper confidence band}
\item{$k:}{[1,1] integer, the input k for reference}
\item{$quantile.fun:}{function, the input quantile.fun for reference}
\item{$wall_time_taken:}{[1,1] double, Time taken to find the result, in seconds}
}
\description{
No multiplicity corrections.
}
