% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{central_quantile}
\alias{central_quantile}
\title{Find central quantiles (densest region of data)}
\usage{
central_quantile(x, k)
}
\arguments{
\item{x}{[1,N] numeric, A 1D data vector}

\item{k}{[1,1] int, Number of values to exclude}
}
\value{
[1,2] numeric, The endpoints [min, max] of the quantile
}
\description{
Symmetrically excludes k extreme values from x, k/2 from each end.
If k is odd, then one less value is excluded from the left tail.
}
