% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{collect_costs}
\alias{collect_costs}
\title{Collect boundary row costs (sub-step of findcb_topdown())}
\usage{
collect_costs(dmat, d.order)
}
\arguments{
\item{dmat}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{d.order}{TODO}
}
\value{
A data.frame with costs for each row of data
}
\description{
Collect costs attainable from rows that lie at the envelope boundary
Uses d.order structure to specify points at the border
}
\keyword{internal}
