% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{createEnvelopeMasks}
\alias{createEnvelopeMasks}
\title{Initial creation of the envelope mask data structure}
\usage{
createEnvelopeMasks(dmat, Z)
}
\arguments{
\item{dmat}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{Z}{[N,M] logical, Output of some mvci method, mvci$Z}
}
\value{
A list with mvci envelope masks.
}
\description{
Create mvci envelope masks
}
\keyword{internal}
