% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{findcb_bottomup_byrow}
\alias{findcb_bottomup_byrow}
\title{Bottom-up method that adds full rows of data}
\usage{
findcb_bottomup_byrow(dmat, k, l, ciSeedInds)
}
\arguments{
\item{dmat}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{k}{[1,1] integer, Number of observations (rows) to remove}

\item{l}{[1,1] integer, Max number of outlier dimensions for a data row that is counted as within MVCI}

\item{ciSeedInds}{[1,J] integer, Row indices of rows that should be used as seed. Typically the median.}
}
\value{
A list describing the multivariate confidence band.
}
\description{
Find MVCI using a bottom up greedy algorithm.
}
