% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{findcb_topdown}
\alias{findcb_topdown}
\title{Greedy top-down multivariate confidence interval (mvci)}
\usage{
findcb_topdown(dmat, K, L, verbose = F)
}
\arguments{
\item{dmat}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{K}{[1,1] integer, Number of observations (rows) to remove}

\item{L}{[1,1] integer, Max number of outlier dimensions for a data row that is counted as within MVCI}

\item{verbose}{logical, Verbosity of output}
}
\value{
A list describing the multivariate confidence band.
}
\description{
Manuscript algorithm 1: Greedy

Algorithm:
 do k times:
   1. solve k = 0, l > 0 to get CB
   2. decide wich row to remove and remove it
 For the N-k row remaining rows, solve k = 0, l > 0 to get CB.

 Output:
   Z [N, M] logical, indicates included points/rows
   upmask, downmask [N, M] logical, indicates positions of CB border, one TRUE per column
   row.inc.idx [1:(N-K)] integer, Indices of rows that are within CB
   row.exc.idx [1:K] integer,  Indices of rows that are outside CB
   L, K Input parameters for documentation

TODO: Tries to process insane inputs such as L>M
}
