% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{fold.profile}
\alias{fold.profile}
\title{Generic B -fold crossvalidation curve}
\usage{
fold.profile(cbfun, cb.contains.fun, data, B, k.max = floor(0.5 * ((B - 1)/B)
  * nrow(data)))
}
\description{
Input:
  cbfun         Function to compute the confidence bands with, interface: cbmat = cbfun(data, k)
  cb.contains.fun     Function to test if vector x belongs to a given CB
  data          (N,M) numeric matrix, N samples, M dimensions/variables. The samples define the
                empirical cdfs of the m variables.
  B             (1,1) integer, How many cross validation folds to use. l = [2,...,N].
  k.max  (1,1) integer, Threshold for ending the computation. When more than 'k.max'
                of the test fold observations are outside confidence band the profile computation
                ends. Can be used to speed up computations when only the beginning of the profile
                curve is of interest.
}
\details{
Output:
  A list with elements:
  profile   (1,N-floor(N/B)) numeric, Crossvalidation based profile of FWER control
}
