% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_min_interval.R
\name{get_ci}
\alias{get_ci}
\title{Naive quantile based multivariate confidence interval (mvci)}
\usage{
get_ci(X, alpha)
}
\arguments{
\item{X}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{alpha}{[1,1] numeric, Percentage of observations to leave outside the interval per dimension}
}
\value{
A list with upper and lower mvci boundary.
TODO: return value differs from that of other methods.
}
\description{
Mininum width quantile (1-a)*100% CI for a data matrix.
Applies smallest_ci() to each column of X.
Does not control FWER over multiple dimensions like mvci_min_width_quantile().
}
