% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{getEnvelope}
\alias{getEnvelope}
\title{Get confidence band (envelope) based on Z matrix}
\usage{
getEnvelope(dmat, Z)
}
\arguments{
\item{dmat}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{Z}{[N,M] logical, Output of some mvci method, mvci$Z}
}
\value{
A list with two vectors indicating values of mvci for each dimension
}
\description{
Not used much since envelope masks are easier to use
}
