% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{getEnvelopeRow}
\alias{getEnvelopeRow}
\title{Get confidence band (envelope) based on row indices}
\usage{
getEnvelopeRow(dmat, rowinds)
}
\arguments{
\item{dmat}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{rowinds}{[] interger, TODO}
}
\value{
A list with two vectors indicating values of mvci for each dimension
}
\description{

}
