% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{getRows}
\alias{getRows}
\title{Identify rows that are inside/outside confidence band based on Z}
\usage{
getRows(Z, type)
}
\arguments{
\item{Z}{[N,M] logical, Output of some mvci method, mvci$Z}

\item{type}{char {'inside', 'outside'}, inside -> rows completely TODO}
}
\value{
[N,1] logical, A logical vector indicating rows that match 'type'
}
\description{
Identify rows that lie completely within the mvci using mvci$Z
}
