% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_profile.R
\name{k.profile}
\alias{k.profile}
\title{Generic test set "k profile" (using whichever CB algorithm)}
\usage{
k.profile(cbfun, cb.contains.fun, data.train, data.test, k.max = floor(0.5 *
  nrow(data.train)))
}
\description{
Determines the iterations k at which observations in the test set become extreme
i.e. do not belong to the confidence band anymore. The result is an (1,N) vector
of integers {0,1,...,nrow(data.test} identifying how many of the test set observations
are extreme at a given k=[1,...,N].
}
\details{
This funcition is needed in the fold.profile() algorithm.

Input:
  cbfun         Function to compute the confidence bands with, interface: cbmat = cbfun(data, k)
  cb.contains.fun     Function to test if vector x belongs to a given CB, TRUE = inside, F = outside
  data.train         (N,M) numeric matrix, N samples, M dimensions/variables. The samples define the
                  empirical cdfs of the m variables.
  data.test       (N.test,M) numeric matrix, Test dataset
  k.max    (1,1) integer, Threshold for ending the computation. When more than 'k.max'
                  of the observations in 'data.test' are outside confidence band the profile
                  computation ends. Can be used to speed up computations when only the beginning of the profile
                  curve is of interest.

Output:
  A list with elements:
  k.profile   (1,N) vector of integers [0,1,...,nrow(data.test)]
}
