% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{max_n}
\alias{max_n}
\title{Find n:th largest value of x}
\usage{
max_n(x, n)
}
\arguments{
\item{x}{[1,N] numeric, Data vector}

\item{n}{[1,1] integer, Up to how far to search}
}
\value{
[1,1] numeric, n:th largest value of x
}
\description{
Find n:th largest value of x
From: http://stackoverflow.com/questions/2453326/fastest-way-to-find-second-third-highest-lowest-value-in-vector-or-column
}
