% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{max_n_set}
\alias{max_n_set}
\title{k largest values of x (an extension of base::max())}
\usage{
max_n_set(x, k)
}
\arguments{
\item{x}{(1,N) numeric, Data vector}

\item{k}{(1,1) integer, Up to how far to search}
}
\value{
[1,J] A data.frame with information on the k largest values of x. See k.min() for details.
}
\description{
Searches for maximum values up to k:th largest value. See k.min() for details.
}
