% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{min_n}
\alias{min_n}
\title{Find n:th smallest value of x}
\usage{
min_n(x, n)
}
\arguments{
\item{x}{[1,N] numeric, Data vector}

\item{n}{[1,1] integer, Up to how far to search}
}
\value{
[1,1] numeric, n:th smallest value of x
}
\description{
Find n:th smallest value of x
}
