% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{min_n_set}
\alias{min_n_set}
\title{k smallest values of x (an extension of base::min())}
\usage{
min_n_set(x, k)
}
\arguments{
\item{x}{(1,N) numeric, Data vector}

\item{k}{(1,1) integer, Up to how far to search}
}
\value{
A data.frame with k rows and columns:
\item{k}{1 for min, 2 for 2nd smallest, etc. ...}
\item{ind}{position within x}
\item{value}{x[ind]}
so that x[res$ind] = res$value are the k smallest values, starting with the smallest.
}
\description{
Searches for minimum values up to k:th smallest value in O(k*N) time.
}
