% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{min_width_quantile}
\alias{min_width_quantile}
\title{Find shortest interval quantiles}
\usage{
min_width_quantile(x, k)
}
\arguments{
\item{x}{[1,N] numeric, A 1D data vector}

\item{k}{[1,1] int, Number of values to exclude}
}
\value{
[1,2] numeric, The endpoints [min, max] of the quantile
}
\description{
Excludes k extreme values from x such that the remaining contiguous range has the
shortest width.

Note: If the distribution of x is uniform, e.g. x = 1:N, or otherwise there are
sevaral shortest intervals then the first is returned, in this example 1:(N-k).
}
