% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{mvci_greedy}
\alias{mvci_greedy}
\title{Greedy heuristics for computing MultiVariate Confidence Intervals (MVCI)
bottom-up
top-down
Greedy methods to find MVCIs
TODO: contains only bottom-up, not the more frequent top-down}
\usage{
mvci_greedy(dmat, k, l, method = "bottomup_byrow")
}
\arguments{
\item{dmat}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{k}{[1,1] integer, Number of observations (rows) to remove}

\item{l}{[1,1] integer, Max number of outlier dimensions for a data row that is counted as within MVCI}

\item{method}{char, Method to select, allowed: {'bottomup_byrow'}}
}
\value{
A list describing the multivariate confidence band.
}
\description{
A wrapper to call different methods with different initializations
}
