% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_min_interval.R
\name{mvci_min_width_quantile}
\alias{mvci_min_width_quantile}
\title{Computes minimum width quantile CI separately for each column of matrix X i.e.
separately for each variable}
\usage{
mvci_min_width_quantile(X, k, l)
}
\description{
Adjusts the number of observations to exclude (i.e. the percentage a in function
f()) such that the total number of rows insice the confidence area mathces n-k
}
