% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_profile.R
\name{mwe.fold.profile}
\alias{mwe.fold.profile}
\title{Minimum width envelope confidence region, FWER controlled, l fold crossvalidation}
\usage{
mwe.fold.profile(data, B, k.max = floor(0.5 * ((B - 1)/B) * nrow(data)))
}
\arguments{
\item{data}{(N,M) numeric matrix, N samples, M dimensions/variables.
The samples define the empirical cdfs of the m variables.}

\item{B}{(1,1) integer, How many cross validation folds to use. l = [2,...,N].}

\item{k.max}{(1,1) integer, Threshold for ending the computation. When more than 'k.max'
of the test fold observations are outside confidence band the profile computation
ends. Can be used to speed up computations when only the beginning of the profile
curve is of interest.}
}
\value{
A list with elements:
\item{profile}{(1,N-floor(N/B)) numeric, Crossvalidation based profile of FWER control}
}
\description{
Using cross validation approach aims at approximating the effective k to be used in the mwe.ci()
function so that FWER is controlled at level alpha. Improves upon mwe.fwer.ci() which for small N
outputs instabile results. Crossvalidation smooths out the instability.
}
