% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_profile.R
\name{mwe.k.profile}
\alias{mwe.k.profile}
\title{Test set "k profile" using Minimum Width Envelope greedy algorithm}
\usage{
mwe.k.profile(data.rs, data.test, k.max)
}
\arguments{
\item{data.rs}{(N,M) numeric matrix, N samples, M dimensions/variables. The samples define the
empirical cdfs of the m variables.}

\item{data.test}{(N.test,M) numeric matrix, Test dataset}

\item{k.max}{(1,1) integer, Threshold for ending the computation. When more than 'k.max'
of the observations in 'data.test' are outside confidence band the profile
computation ends. Can be used to speed up computations when only the beginning of the profile
curve is of interest.}
}
\value{
A list with elements:
\item{k.profile}{(1,N) vector of integers [0,1,...,nrow(data.test)]}
}
\description{
Determines the iterations k at which observations in the test set become extreme
i.e. do not belong to the confidence band anymore. The result is an (1,N) vector
of integers {0,1,...,nrow(data.test} identifying how many of the test set observations
are extreme at a given k=[1,...,N].

This funcition is needed in the mwe.fold() algorithm.
}
