% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot.data.cb}
\alias{plot.data.cb}
\title{Plot CB}
\usage{
\method{plot}{data.cb}(dmat, cb, timevec = 1:ncol(dmat), title = "CB plot",
  xlab = "time / variables", ylab = "value", ylim = c(min(dmat),
  max(dmat)), xlim = c(timevec[1], timevec[length(timevec)]),
  draw.cb.points = T, highlight.rows.out = T, samp.frac = 1,
  yintercept = NULL, xaxt = "s")
}
\description{
Plot CB
}
