% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{plot.data.ci}
\alias{plot.data.ci}
\title{Plot CB when also whole rows have been removed}
\usage{
\method{plot}{data.ci}(dmat, Z, upmask, downmask, type = "outside",
  title = "CI plot", xlab = "time / variables", ylab = "value")
}
\description{
Plot CB when also whole rows have been removed
}
