% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot.data.ci2}
\alias{plot.data.ci2}
\title{Plot CB if only points have been removed}
\usage{
\method{plot}{data.ci2}(dmat, upmask, downmask, timevec = 1:ncol(dmat),
  title = "CI plot", xlab = "time / variables", ylab = "value",
  ylim = c(min(dmat), max(dmat)), draw.cb.points = T, samp.frac = 1,
  yintercept = NULL)
}
\description{
Plot CB if only points have been removed
}
