% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_data_cb}
\alias{plot_data_cb}
\title{Plot CB}
\usage{
plot_data_cb(dmat, cb, timevec = 1:ncol(dmat), title = "CB plot",
  xlab = "time / variables", ylab = "value", ylim = c(min(dmat),
  max(dmat)), xlim = c(timevec[1], timevec[length(timevec)]),
  draw.cb.points = T, highlight.rows.out = T, samp.frac = 1,
  yintercept = NULL, xaxt = "s")
}
\arguments{
\item{dmat}{[N,M] numeric, Data matrix}

\item{cb}{list, The output of some mvci method}

\item{timevec}{[1,M] numeric, Plot positions of the columns of dmat}

\item{title}{char, Plot title}

\item{xlab}{char, Plot x axis label}

\item{ylab}{char, Plot y axis label}

\item{ylim}{[1,2] numeric, y axis limits}

\item{xlim}{[1,2] numeric, x axis limits}

\item{draw.cb.points}{[1,1] logical, Draw points that define CB boundary}

\item{highlight.rows.out}{[1,1] logical, Should curves that exit CB be highlighted?}

\item{samp.frac}{[1,1] numeric, Sampling fraction from range [0,1]}

\item{yintercept}{[1,1] numeric, Horizontal line y axis intercept, default: no horizontal line}

\item{xaxt}{char, Passed on to base::plot()}
}
\value{
Base graphics plot to current plotting device
}
\description{
Plot CB
}
