% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_data_ci}
\alias{plot_data_ci}
\title{Plot CB when also whole rows have been removed}
\usage{
plot_data_ci(dmat, Z, upmask, downmask, type = "outside", title = "CI plot",
  xlab = "time / variables", ylab = "value")
}
\arguments{
\item{dmat}{[N,M] numeric, Data matrix}

\item{Z}{[N,M] logical, Output of some mvci method, mvci$Z}

\item{upmask}{[N,M] logical, Output of some mvci method, mvci$upmask}

\item{downmask}{[N,M] logical, Output of some mvci method, mvci$downmask}

\item{type}{char {'inside','outside'}, 'outside' -> plot rows outside CB in red, 'inside' -> plot rows inside CB in red}

\item{title}{char, Plot title}

\item{xlab}{char, Plot x axis label}

\item{ylab}{char, Plot y axis label}
}
\value{
Base graphics plot to current plotting device
}
\description{
Plot CB when also whole rows have been removed
}
