% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_data_ci2}
\alias{plot_data_ci2}
\title{Plot CB if only points have been removed}
\usage{
plot_data_ci2(dmat, upmask, downmask, timevec = 1:ncol(dmat),
  title = "CI plot", xlab = "time / variables", ylab = "value",
  ylim = c(min(dmat), max(dmat)), draw.cb.points = T, samp.frac = 1,
  yintercept = NULL)
}
\arguments{
\item{dmat}{[N,M] numeric, Data matrix}

\item{upmask}{[N,M] logical, Output of some mvci method, mvci$upmask}

\item{downmask}{[N,M] logical, Output of some mvci method, mvci$downmask}

\item{timevec}{[1,M] numeric, Plot positions of the columns of dmat}

\item{title}{char, Plot title}

\item{xlab}{char, Plot x axis label}

\item{ylab}{char, Plot y axis label}

\item{ylim}{[1,2] numeric, y axis limits}

\item{draw.cb.points}{[1,1] logical, Draw points that define CB boundary}

\item{samp.frac}{[1,1] numeric, Sampling fraction from range [0,1]}

\item{yintercept}{[1,1] numeric, Horizontal line y axis intercept, default: no horizontal line}
}
\value{
Base graphics plot to current plotting device
}
\description{
Plot CB if only points have been removed
}
