% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{points_outside_ci}
\alias{points_outside_ci}
\title{Find points on a row that lie outside confidence band}
\usage{
points_outside_ci(rowind, dmat, mvci)
}
\arguments{
\item{rowind}{integer, A row index}

\item{dmat}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{mvci}{list, mvci result list}
}
\value{
A data.frame with colum indices, $isover and cost
}
\description{
Report each point along with associated costs
}
\keyword{internal}
