% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{remove_points_cb}
\alias{remove_points_cb}
\title{Minimize envelope for k=0, l>0 (sub-step of findcb_topdown())}
\usage{
remove_points_cb(dmat, l)
}
\arguments{
\item{dmat}{[N,M] numeric, Data, vector valued M-dimensional observations on rows,}

\item{l}{[1,1] integer, Max number of outlier dimensions for a data row that is counted as within MVCI}
}
\value{
A list describing the multivariate confidence band.
}
\description{
Solve problem k = 0, l > 0 by greedily removing points starting from the
whole data envelope.
To be used as a substep of the greedy top-down approach
Manuscript algorithm 2: FindEnvelope(X, I, L)
}
\keyword{internal}
