% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{repmat}
\alias{repmat}
\title{Create a large matrix from copies of a small one}
\usage{
repmat(X, i, k)
}
\arguments{
\item{X}{[a,b] numeric, A matrix}

\item{i}{[1,1] integer, Number of repeats over rows}

\item{k}{[1,1] integer, Number of repeats over cols}
}
\value{
[a*n, b*m] numeric, A matrix created by repeating X
}
\description{
Functions like repmat() in Matlab:
http://waxworksmath.com/Authors/G_M/Hastie/Code/Chapter4/repmat.R
}
\examples{
repmat(matrix(1:4, nrow = 2, ncol = 2, byrow = TRUE), 4, 3)

}
