% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{row.outside.ci}
\alias{row.outside.ci}
\title{Find observations that are outside confidence intervals in at least one variable}
\usage{
row.outside.ci(data.train, cb.low, cb.high)
}
\arguments{
\item{data.train}{(n,m) numeric matrix, n samples, m variables. The samples define the empirical cdfs of the m variables. Variables correspond to multiple hypotheses.}

\item{cb.low}{(1,m) numeric vector, lower CI limit for each of the dimensions}

\item{cb.high}{(1,m) numeric vector, upper CI limit for each of the dimensions}
}
\value{
[M,1] logical, True if the observation (row) exits the mvci
}
\description{
Find observations that are outside confidence intervals in at least one variable
}
