% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{run.exp1}
\alias{run.exp1}
\title{Experiment 1: Compute Pr[V(x|x_u,x_l)<=L] for various value of L, N_train and B}
\usage{
run.exp1(cbfun, X_train, X_test, N.train.arr, M.arr, K.prc, L.arr, B)
}
\description{
Experiment 1: Compute Pr[V(x|x_u,x_l)<=L] for various value of L, N_train and B
}
