% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_min_interval.R
\name{smallest_ci}
\alias{smallest_ci}
\title{Mininum width quantile CI for a vector of data
Returns (1 - k/length(x)) * 100% CI for a data vector using minimum width quantiles}
\usage{
smallest_ci(x, k)
}
\description{
todo: same as min_width_quantile()
}
