% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{total.contained}
\alias{total.contained}
\title{Total number of rows in X that do not violate cb}
\usage{
total.contained(X, cb.low, cb.high, L)
}
\description{
Total number of rows in X that do not violate cb
}
