% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_min_interval.R
\name{total_contained}
\alias{total_contained}
\title{Returns the number of rows in matrix X that are within ci with respect
to relaxation l}
\usage{
total_contained(ci, X, l)
}
\description{
Returns the number of rows in matrix X that are within ci with respect
to relaxation l
}
