% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci.R
\name{updateEnvelopeMasks}
\alias{updateEnvelopeMasks}
\title{Update mvci envelope mask}
\usage{
updateEnvelopeMasks(mvci, rowInd, pointDF)
}
\arguments{
\item{mvci}{list, mvci result list}

\item{rowInd}{integer, Index of the row to add}

\item{pointDF}{data.frame, Data point data.frame}
}
\value{
An updated mvci list
}
\description{
Update the mvci envelope mask data structure by adding a new row
}
\keyword{internal}
