% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_min_interval.R
\name{ci_contains}
\alias{ci_contains}
\title{Returns TRUE if ci contains x with respect to relaxation l and FALSE otherwise}
\usage{
ci_contains(ci, x, l)
}
\description{
Returns TRUE if ci contains x with respect to relaxation l and FALSE otherwise
}
