% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_min_interval.R
\name{ci_violations}
\alias{ci_violations}
\title{Returns the number of points in which x lies outside ci}
\usage{
ci_violations(ci, x)
}
\description{
Returns the number of points in which x lies outside ci
}
