% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvci_tools.R
\name{prc.inside.cb}
\alias{prc.inside.cb}
\title{Compute Pr[V(x|x_u,x_l)<=L] using X_train, K, L and X_test}
\usage{
prc.inside.cb(cbfun, X_train, X_test, L)
}
\description{
Compute Pr[V(x|x_u,x_l)<=L] using X_train, K, L and X_test
}
